trackRai: Multi-Object Tracking With YOLO11

trackRai is a YOLO-based multi-object tracking software for R. It provides a series of Shiny apps for automating the preparation and training of YOLO11 models, for performing the tracking of multiple objects in a video, and for visualizing the results of the tracking process.

See also

Author

Maintainer: Simon Garnier garnier@njit.edu (ORCID)

Other contributors:

National Science Foundation, Award ID 2222418 [funder]